Use sparse-checkout add instead of set
By using set you reset the entire sparse-checkout configuration,
including the default root directory (/* !/*/) which we don't want
to remove.

Added some extra command logging so we can easily repro locally without
needing to find the code that ran.

Update to "/eng" so that we only get paths that start with eng instead
of every path that has eng as a folder.